Set logdir to work with new dhcp logging patch
Started having failures of Instances to boot in tempest runs, turns out
it's a side-effect of this patch here:
    https://github.com/openstack-dev/devstack/commit/ac8ff0f1e9e5116c390bf8a711c950bcaca71d32

We were setting SCREEN_LOGDIR=/opt/stack/screen-logs, but this patch hard codes a check for
the dhcpbridge log in /opt/stack/logs.  This change just modifies our setup a bit so that
everything just gets dumped to /opt/stack/logs and fixes the problem that way.