Move to yaml for config
We decided that using "run <CI-NAME>" was a good standard to use
for rechecks that wouldn't impact the official gates.

I've never liked the iniparse module that I was using, and it proved
to be cumbersome to deal with white-space in strings, so I finally
just nuked it and switched to using yaml.

So now we have a yaml formatted config, and I added the "run ci-name"
in the sample config.yaml file.