Initial commit supporting active-active c-vol
This change includes support for running a pair of cinder volume services on
separate nova VM instances for the purpose of executing temptest tests against
an active-active cinder volume service setup.

This does not slot into the existing sos-ci configuration, this is simply a
means for sharing the new ansible playbooks to solicit feedback.