Check for Depends-On in commit message
Add a function to check for Depends-On in the commit message of changes
and query gerrit for information about the dependencies.

This will allow the executor to build a ZUUL_CHANGES environment variable
that can be passed to the test runs like zuul does it in openstack-infra.

Note: Most of the code is stolen from the openstack-infra/zuul project.

While at it, refactor the SSH connection code into a dedicated module,
allowing for better reuse.